pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "gymnasium>=1.2.1",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
]
